+++ buession-jdbc/src/main/java/com/buession/jdbc/datasource/HikariDataSource.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.jdbc.datasource;
@@ -56,11 +56,7 @@ public class HikariDataSource extends AbstractDataSource<com.zaxxer.hikari.Hikar
 
 	@Override
 	public com.zaxxer.hikari.HikariDataSource createDataSource(){
-		com.zaxxer.hikari.HikariDataSource dataSource = new com.zaxxer.hikari.HikariDataSource();
-
-		applyPoolConfiguration(dataSource, getPoolConfiguration());
-
-		return dataSource;
+		return doCreateDataSource(new com.zaxxer.hikari.HikariDataSource());
 	}
 
 	@Override